GBAY-154 int to long

+++ DataAccess/Interfaces/ISubGroupRepository.cs
@@ -4,14 +4,14 @@ namespace dotnet.DataAccess.Interfaces
 {
 	public interface ISubGroupRepository
 	{
-		Task<SubGroup> GetByIdAsync(int id);
+		Task<SubGroup> GetByIdAsync(long id);
 		Task<IEnumerable<SubGroup>> GetAllAsync();
 		Task AddAsync(SubGroup subGroup);
 		Task UpdateAsync(SubGroup subGroup);
-		Task DeleteAsync(int id);
+		Task DeleteAsync(long id);
 
 		// Additional get methods
 		Task<SubGroup?> GetByNameAsync(string name);
-		Task<IEnumerable<SubGroup>> GetByMainGroupIdAsync(int mainGroupId);
+		Task<IEnumerable<SubGroup>> GetByMainGroupIdAsync(long mainGroupId);
 	}
 }
